refactor coherence tests into test_gdal_python
